:recycle: Split up users schema
