+++ creator/users/schema.py
@@ -1,32 +1,24 @@
 import graphene
 import django_filters
-from graphene import (
-    relay,
-    Connection,
-    ObjectType,
-    Field,
-    List,
-    String,
-    DateTime,
-)
+from graphene import relay
 from graphql_relay import from_global_id
+from graphql import GraphQLError
 from graphene_django import DjangoObjectType
-from graphene_django.utils import get_model_fields
 from graphene_django.converter import convert_django_field_with_choices
 from graphene_django.filter import DjangoFilterConnectionField
 from django_filters import OrderingFilter
 
-from graphql import GraphQLError
 from django.contrib.auth import get_user_model
 from django.contrib.auth.models import Group, Permission
 
-from creator.studies.models import Study, Membership, MEMBER_ROLE_CHOICES
-from creator.studies.schema import StudyNode
+from creator.studies.models import Study, Membership
+
+from creator.users.mutations import Mutation
 
 User = get_user_model()
 
 
-class CollaboratorConnection(Connection):
+class CollaboratorConnection(graphene.Connection):
     class Meta:
         abstract = True
 
@@ -37,8 +29,8 @@ class CollaboratorConnection(Connection):
         """
 
         # Need to referrence the node by module to avoid circular deps
-        invited_by = Field("creator.users.schema.UserNode")
-        joined_on = DateTime()
+        invited_by = graphene.Field("creator.users.schema.UserNode")
+        joined_on = graphene.DateTime()
         # We need to manually convert the role field's choices into enums
         # that graphene understands
         role = convert_django_field_with_choices(
@@ -99,7 +91,9 @@ class CollaboratorConnection(Connection):
 
 
 class UserNode(DjangoObjectType):
-    roles = List(String, description="Roles that the user has")
+    roles = graphene.List(
+        graphene.String, description="Roles that the user has"
+    )
 
     def resolve_roles(self, info):
         return self.ego_roles
@@ -224,150 +218,11 @@ class PermissionFilter(django_filters.FilterSet):
         fields = {"name": ["exact"], "codename": ["exact"]}
 
 
-class MyProfileMutation(graphene.Mutation):
-    """
-    Updates user's profile
-    Only modifies the request's authenticated user's profile.
-    We can only change fields that originate in our data model as the other
-    fields come from our authentication services.
-    """
-
-    class Arguments:
-        slack_notify = graphene.Boolean()
-        slack_member_id = graphene.String()
-        email_notify = graphene.Boolean()
-
-    user = graphene.Field(UserNode)
-
-    def mutate(self, info, **kwargs):
-        """
-        Updates a user's profile.
-        Only applies to the request's current user
-        """
-        user = info.context.user
-        if not user.is_authenticated or user is None or user.email == "":
-            raise GraphQLError("Not authenticated to mutate profile")
-
-        if "slack_notify" in kwargs:
-            user.slack_notify = kwargs.get("slack_notify")
-        if "slack_member_id" in kwargs:
-            user.slack_member_id = kwargs.get("slack_member_id")
-        if "email_notify" in kwargs:
-            user.email_notify = kwargs.get("email_notify")
-        user.save()
-
-        return MyProfileMutation(user=user)
-
-
-class UpdateUserMutation(graphene.Mutation):
-    """
-    Updates user's groups
-    """
-
-    class Arguments:
-        user = graphene.ID(required=True)
-        groups = graphene.List(graphene.ID, required=False)
-
-    user = graphene.Field(UserNode)
-
-    def mutate(self, info, user, groups=None):
-        """
-        Updates a user
-        """
-        current_user = info.context.user
-        if not current_user.has_perm("creator.change_user"):
-            raise GraphQLError("Not allowed")
-
-        _, user_id = from_global_id(user)
-        try:
-            user = User.objects.get(id=user_id)
-        except User.DoesNotExist:
-            raise GraphQLError("User does not exist.")
-
-        if isinstance(groups, list):
-            group_ids = [from_global_id(g)[1] for g in groups]
-            group_objs = Group.objects.filter(id__in=group_ids).all()
-            # If fewer groups were returned from the database than
-            # were requested, one or more must not exist
-            if len(group_objs) < len(groups):
-                raise GraphQLError("Group does not exist.")
-            user.groups.set(group_objs, clear=True)
-
-        return UpdateUserMutation(user=user)
-
-
-class SubscribeToMutation(graphene.Mutation):
-    """
-    Subscribe a user to a study
-    admin - may subscribe to any study
-    service - may not subscribe to studies
-    user - may only subscribe to studies which they are in the group of
-    unauthed - may not subscribe to studies
-    """
-
-    class Arguments:
-        study_id = graphene.String(required=True)
-
-    success = graphene.Boolean()
-    user = graphene.Field(UserNode)
-
-    def mutate(self, info, study_id, **kwargs):
-        """
-        Adds a study to the user's study subscriptions
-        """
-        user = info.context.user
-        if user is None or not user.is_authenticated:
-            raise GraphQLError("Not authenticated to subscribe")
-
-        try:
-            study = Study.objects.get(kf_id=study_id)
-        except Study.DoesNotExist:
-            raise GraphQLError("Study does not exist.")
-
-        if study_id not in user.ego_groups and "ADMIN" not in user.ego_roles:
-            raise GraphQLError("Not authenticated to subscribe")
-
-        # Add the study to the users subscriptions
-        user.study_subscriptions.add(study)
-
-        return SubscribeToMutation(success=True, user=user)
-
-
-class UnsubscribeFromMutation(graphene.Mutation):
-    """
-    Unsubscribes a user from a study
-    """
-
-    class Arguments:
-        study_id = graphene.String(required=True)
-
-    success = graphene.Boolean()
-    user = graphene.Field(UserNode)
-
-    def mutate(self, info, study_id, **kwargs):
-        """
-        Removes a study from the user's study subscriptions
-        """
-        user = info.context.user
-        if user is None or not user.is_authenticated:
-            raise GraphQLError("Not authenticated to unsubscribe")
-
-        try:
-            study = Study.objects.get(kf_id=study_id)
-        except Study.DoesNotExist:
-            raise GraphQLError("Study does not exist.")
-
-        # Remove the study to the users subscriptions
-        user.study_subscriptions.remove(study)
-
-        return SubscribeToMutation(success=True, user=user)
-
-
 class Query(object):
     all_users = DjangoFilterConnectionField(
         UserNode, filterset_class=UserFilter
     )
-    my_profile = Field(UserNode)
+    my_profile = graphene.Field(UserNode)
     group = relay.Node.Field(GroupNode, description="Get a group")
     all_groups = DjangoFilterConnectionField(
         GroupNode, filterset_class=GroupFilter, description="List all groups"
